chore: Remove multi tenancy
To simplify the application we remove the multi tenancy concern, since we only have a single tenant and don't expect to have more soon.
