--- src/Services/CRM/Application/Contracts/Services/ITenantService.cs
@@ -1,8 +1,0 @@
-﻿using System;
-namespace KDVManager.Services.CRM.Application.Contracts.Services
-{
-    public interface ITenantService
-    {
-        Guid Tenant { get; }
-    }
-}
