--- src/Services/CRM/Application/Exceptions/TenantRequiredException.cs
@@ -1,10 +1,0 @@
-﻿using System;
-
-namespace KDVManager.Services.CRM.Application.Exceptions;
-
-public class TenantRequiredException : ApplicationException
-{
-    public TenantRequiredException() : base($"Tenant claim required.")
-    {
-    }
-}
