--- src/Services/CRM/Domain/Entities/Tenant.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace KDVManager.Services.CRM.Domain.Entities
-{
-    public class Tenant
-    {
-        public Guid Id { get; set; }
-    }
-}
