--- src/Services/CRM/Domain/Interfaces/IMustHaveTenant.cs
@@ -1,9 +1,0 @@
-using System;
-
-namespace KDVManager.Services.CRM.Domain.Interfaces
-{
-    public interface IMustHaveTenant
-    {
-        Guid TenantId { get; set; }
-    }
-}
