+++ src/Services/CRM/Infrastructure/CRMDbContext.cs
@@ -1,0 +1,21 @@
+﻿using KDVManager.Services.CRM.Domain.Entities;
+using Microsoft.EntityFrameworkCore;
+using System.Threading.Tasks;
+using System.Threading;
+
+namespace KDVManager.Services.CRM.Infrastructure
+{
+    public class CRMDbContext : DbContext
+    {
+        public CRMDbContext(DbContextOptions<CRMDbContext> options) : base(options)
+        { }
+
+        public DbSet<Child> Children { get; set; }
+        public DbSet<Person> People { get; set; }
+
+        protected override void OnModelCreating(ModelBuilder modelBuilder)
+        {
+            base.OnModelCreating(modelBuilder);
+        }
+    }
+}
