+++ src/Services/CRM/Infrastructure/Configurations/ChildConfiguration.cs
@@ -14,9 +14,6 @@ namespace KDVManager.Services.CRM.Infrastructure.Configurations
         {
             builder.Property(e => e.FamilyName)
                 .IsRequired();
-
-            builder.Property(e => e.TenantId)
-                .IsRequired();
         }
     }
 }
