+++ src/Services/Scheduling/Api/ConfigureServices.cs
@@ -1,5 +1,3 @@
-using KDVManager.Services.Scheduling.Api.Services;
-using KDVManager.Services.Scheduling.Application.Contracts.Services;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.OpenApi.Models;
 using Microsoft.OpenApi.Any;
@@ -47,8 +45,6 @@ public static class ConfigureServices
                 options.Authority = domain;
                 options.Audience = configuration["Auth0:Audience"];
             });
-
-        services.AddScoped<ITenantService, TenantService>();
 
         return services;
     }
