+++ src/Services/Scheduling/Application/Contracts/Services/ITenantService.cs
@@ -1,7 +1,1 @@
-﻿using System;
-namespace KDVManager.Services.Scheduling.Application.Contracts.Services;
-
-public interface ITenantService
-{
-    Guid Tenant { get; }
-}
+﻿// This file has been removed as per the suggested code change.
