+++ src/Services/Scheduling/Domain/Entities/Group.cs
@@ -1,13 +1,10 @@
 ﻿using System;
-using KDVManager.Services.Scheduling.Domain.Interfaces;
 
 namespace KDVManager.Services.Scheduling.Domain.Entities;
 
-public class Group : IMustHaveTenant
+public class Group
 {
     public Guid Id { get; set; }
-
-    public Guid TenantId { get; set; }
 
     public string Name { get; set; }
