+++ src/Services/Scheduling/Domain/Entities/Schedule.cs
@@ -1,13 +1,11 @@
 using System;
 using System.Collections.Generic;
-using KDVManager.Services.Scheduling.Domain.Interfaces;
 
 namespace KDVManager.Services.Scheduling.Domain.Entities;
 
-public class Schedule : IMustHaveTenant
+public class Schedule
 {
     public Guid Id { get; set; }
-    public Guid TenantId { get; set; }
     public Guid ChildId { get; set; }
     public Guid GroupId { get; set; }
     public DateTime StartDate { get; set; }
