+++ src/Services/Scheduling/Domain/Entities/ScheduleRule.cs
@@ -1,13 +1,11 @@
 using System;
 using System.ComponentModel.DataAnnotations;
-using KDVManager.Services.Scheduling.Domain.Interfaces;
 
 namespace KDVManager.Services.Scheduling.Domain.Entities;
 
-public class ScheduleRule : IMustHaveTenant
+public class ScheduleRule
 {
     public Guid Id { get; set; }
-    public Guid TenantId { get; set; }
     public DayOfWeek Day { get; set; }
     public Guid ScheduleId { get; set; }
     [Required]
