+++ src/Services/Scheduling/Domain/Entities/TimeSlot.cs
@@ -1,17 +1,11 @@
 ﻿using System;
-using KDVManager.Services.Scheduling.Domain.Interfaces;
 
 namespace KDVManager.Services.Scheduling.Domain.Entities;
 
-public class TimeSlot : IMustHaveTenant
+public class TimeSlot
 {
     public Guid Id { get; set; }
-
-    public Guid TenantId { get; set; }
-
     public string Name { get; set; }
-
     public TimeOnly StartTime { get; set; }
-
     public TimeOnly EndTime { get; set; }
 }
