+++ src/Services/Scheduling/Domain/Interfaces/IMustHaveTenant.cs
@@ -1,8 +1,3 @@
 using System;
 
 namespace KDVManager.Services.Scheduling.Domain.Interfaces;
-
-public interface IMustHaveTenant
-{
-    Guid TenantId { get; set; }
-}
