--- src/Services/Scheduling/Infrastructure/MigrationDbContext.cs
@@ -1,24 +1,0 @@
-﻿using KDVManager.Services.Scheduling.Domain.Entities;
-using Microsoft.EntityFrameworkCore;
-using System.Reflection;
-
-namespace KDVManager.Services.Scheduling.Infrastructure;
-
-public class MigrationDbContext : DbContext
-{
-    public MigrationDbContext(DbContextOptions<MigrationDbContext> options) : base(options)
-    {
-    }
-
-    public DbSet<Group> Groups { get; set; }
-    public DbSet<TimeSlot> TimeSlots { get; set; }
-    public DbSet<Schedule> Schedules { get; set; }
-    public DbSet<ScheduleRule> ScheduleRules { get; set; }
-
-    protected override void OnModelCreating(ModelBuilder modelBuilder)
-    {
-        base.OnModelCreating(modelBuilder);
-        modelBuilder.ApplyConfigurationsFromAssembly(Assembly.GetExecutingAssembly());
-    }
-}
-
