+++ src/Services/Scheduling/Infrastructure/SchedulingDbContext.cs
@@ -1,19 +1,13 @@
 ﻿using KDVManager.Services.Scheduling.Domain.Entities;
-using KDVManager.Services.Scheduling.Application.Contracts.Services;
 using Microsoft.EntityFrameworkCore;
-using System.Threading.Tasks;
-using System.Threading;
-using KDVManager.Services.Scheduling.Domain.Interfaces;
 using System.Reflection;
 
 namespace KDVManager.Services.Scheduling.Infrastructure;
 
 public class SchedulingDbContext : DbContext
 {
-    public ITenantService _tenantService;
-    public SchedulingDbContext(DbContextOptions<SchedulingDbContext> options, ITenantService tenantService) : base(options)
+    public SchedulingDbContext(DbContextOptions<SchedulingDbContext> options) : base(options)
     {
-        _tenantService = tenantService;
     }
 
     public DbSet<Group> Groups { get; set; }
@@ -24,27 +18,8 @@ public class SchedulingDbContext : DbContext
     protected override void OnModelCreating(ModelBuilder modelBuilder)
     {
         base.OnModelCreating(modelBuilder);
-        modelBuilder.Entity<Group>().HasQueryFilter(a => a.TenantId == _tenantService.Tenant);
-        modelBuilder.Entity<TimeSlot>().HasQueryFilter(a => a.TenantId == _tenantService.Tenant);
-        modelBuilder.Entity<ScheduleRule>().HasQueryFilter(a => a.TenantId == _tenantService.Tenant);
-        modelBuilder.Entity<Schedule>().HasQueryFilter(a => a.TenantId == _tenantService.Tenant).HasMany(si => si.ScheduleRules);
+        modelBuilder.Entity<Schedule>().HasMany(si => si.ScheduleRules);
 
         modelBuilder.ApplyConfigurationsFromAssembly(Assembly.GetExecutingAssembly());
-    }
-
-    public override async Task<int> SaveChangesAsync(CancellationToken cancellationToken = new CancellationToken())
-    {
-        foreach (var entry in ChangeTracker.Entries<IMustHaveTenant>())
-        {
-            switch (entry.State)
-            {
-                case EntityState.Added:
-                case EntityState.Modified:
-                    entry.Entity.TenantId = _tenantService.Tenant;
-                    break;
-            }
-        }
-        var result = await base.SaveChangesAsync(cancellationToken);
-        return result;
     }
 }
