+++ src/web/package.json
@@ -4,7 +4,7 @@
   "private": true,
   "engines": {
     "node": "20.x",
-    "pnpm": "9.x"
+    "pnpm": "10.x"
   },
   "dependencies": {
     "@auth0/auth0-react": "^2.2.4",
